+++ cli/logs.py
@@ -1,8 +1,10 @@
 import click
-from core.helper import login_required, get, download_log_file, local_only
+from core.helper import (login_required, get, download_log_file,
+                         local_only, download_dump)
 from core.print_formatters import print_logs
 
 from configs.cli_logger import LOG_FILEPATH, DEBUG_LOG_FILEPATH
+
 
 @click.group()
 def logs_cli():
@@ -43,3 +45,35 @@ def cli(debug):
     filepath = DEBUG_LOG_FILEPATH if debug else LOG_FILEPATH
     with open(filepath, 'r') as fin:
         print(fin.read())
+
+
+@logs.command(help="Download log file from container on the connected node")
+@click.argument('name')
+@click.option(
+    '--lines',
+    '-l',
+    help='Output specified number of lines at the end of logs',
+    default=None
+)
+@login_required
+def container(name, lines):
+    params = {'container_name': name}
+    if lines:
+        params['lines'] = lines
+    container_logs = get('container_logs', params)
+    print(container_logs)
+
+
+@logs.command(help="Dump all logs from the connected node")
+@click.option(
+    '--container',
+    '-c',
+    help='Dump logs only from specified container',
+    default=None
+)
+@click.argument('path')
+@login_required
+def dump(container, path):
+    res = download_dump(path, container)
+    if res:
+        print(f'File {res} downloaded')