+++ configs/cli_logger.py
@@ -1,5 +1,5 @@
 import os
-from configs.node import NODE_DATA_PATH
+from configs.node import HOME_DIR
 
 LOG_FORMAT = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
 
@@ -7,8 +7,7 @@ LOG_FILE_SIZE_MB = 300
 LOG_FILE_SIZE_BYTES = LOG_FILE_SIZE_MB * 1000000
 
 LOG_BACKUP_COUNT = 1
-
-LOG_DATA_PATH = os.path.join(NODE_DATA_PATH, 'log')
+LOG_DATA_PATH = os.path.join(HOME_DIR, '.skale-cli-log')
 
 LOG_FILENAME = 'node-cli.log'
 DEBUG_LOG_FILENAME = 'debug-node-cli.log'