+++ core/user.py
@@ -35,7 +35,7 @@ def register_user(config, username, password, token):
 def login_user(config, username, password):
     host = safe_get_config(config, 'host')
     if not host:
-        return
+        host = get_localhost_endpoint()
 
     data = {
         'username': username,
@@ -77,7 +77,7 @@ def show_registration_token(short):
             print(config["token"])
         else:
             print(f'User registration token: {config["token"]}')
-    except FileNotFoundError as e:
+    except FileNotFoundError:
         err_msg = "Couldn't find registration tokens file. Check that node inited on this machine."
         logger.error(err_msg)
         print(err_msg)